+++ assets/airports/airportLoadList.js
@@ -58,6 +58,13 @@ window.AIRPORT_LOAD_LIST = (function() { // eslint-disable-line wrap-iife
      */
     return [
         {
+            icao: 'cyhz',
+            level: 'beginner',
+            name: 'Halifax Stanfield International Airport',
+            premium: false,
+            wip: false
+        },
+        {
             icao: 'cyow',
             level: 'easy',
             name: 'Ottawa Macdonald-Cartier International Airport',