+++ test/trafficGenerator/SpawnPatternModel.spec.js
@@ -15,10 +15,10 @@ import {
     DEPARTURE_PATTERN_MOCK,
     DEPARTURE_PATTERN_ROUTE_STRING_MOCK,
     ARRIVAL_PATTERN_MOCK,
+    ARRIVAL_PATTERN_MOCK_ALL_STRINGS,
     ARRIVAL_PATTERN_CYCLIC_MOCK,
     ARRIVAL_PATTERN_WAVE_MOCK,
-    ARRIVAL_PATTERN_ROUTE_STRING_MOCK,
-    ARRIVAL_PATTERN_SINGLE_ENTRY_AND_RWY_MOCK
+    ARRIVAL_PATTERN_ROUTE_STRING_MOCK
 } from './_mocks/spawnPatternMocks';
 import { INVALID_NUMBER } from '../../src/assets/scripts/client/constants/globalConstants';
 import { DEFAULT_SCREEN_POSITION } from '../../src/assets/scripts/client/constants/positionConstants';
@@ -54,6 +54,15 @@ ava('does not throw when called with valid parameters', (t) => {
     t.notThrows(() => new SpawnPatternModel(ARRIVAL_PATTERN_ROUTE_STRING_MOCK));
 });
 
+ava('initializes correctly when spawn pattern definition uses string type for number values', (t) => {
+    const model = new SpawnPatternModel(ARRIVAL_PATTERN_MOCK_ALL_STRINGS);
+
+    t.true(model._minimumAltitude === 36000);
+    t.true(model._maximumAltitude === 36000);
+    t.true(model.speed === 320);
+    t.true(model.rate === 10);
+});
+
 ava('#position defaults to DEFAULT_SCREEN_POSITION', (t) => {
     const model = new SpawnPatternModel(DEPARTURE_PATTERN_MOCK);
 
@@ -66,6 +75,7 @@ ava('#altitude returns a random altitude rounded to the nearest 1,000ft', (t) =>
     const expectedResult = _round(result, -3);
 
     t.true(_isEqual(result, expectedResult));
+    t.true(typeof result === 'number');
 });
 
 ava('.cycleStart() returns early if cycleStartTime does not equal -1', (t) => {
@@ -156,7 +166,7 @@ ava.skip('._calculateNextWaveDelayPeriod()', (t) => {
     // t.true(result === 360);
 });
 
-ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when an array is passed ', (t) => {
+ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when an array of numbers is passed ', (t) => {
     // creating new mock here so as not to overwrite and affect original
     const arrivalMock = Object.assign({}, ARRIVAL_PATTERN_MOCK, { altitude: 0 });
     const altitudeMock = [10000, 20000];
@@ -166,6 +176,18 @@ ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when an ar
 
     t.true(model._minimumAltitude === altitudeMock[0]);
     t.true(model._maximumAltitude === altitudeMock[1]);
+});
+
+ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when an array of strings is passed ', (t) => {
+    // creating new mock here so as not to overwrite and affect original
+    const arrivalMock = Object.assign({}, ARRIVAL_PATTERN_MOCK, { altitude: 0 });
+    const altitudeMock = ['10000', '20000'];
+    const model = new SpawnPatternModel(arrivalMock);
+
+    model._setMinMaxAltitude(altitudeMock);
+
+    t.true(model._minimumAltitude === 10000);
+    t.true(model._maximumAltitude === 20000);
 });
 
 ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when a number is passed ', (t) => {
@@ -178,6 +200,18 @@ ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when a num
 
     t.true(model._minimumAltitude === altitudeMock);
     t.true(model._maximumAltitude === altitudeMock);
+});
+
+ava('._setMinMaxAltitude() sets _minimumAltitude and _maximumAltitude when a string is passed ', (t) => {
+    // creating new mock here so as not to overwrite and affect original
+    const arrivalMock = Object.assign({}, ARRIVAL_PATTERN_MOCK, { altitude: 0 });
+    const altitudeMock = '23000';
+    const model = new SpawnPatternModel(arrivalMock);
+
+    model._setMinMaxAltitude(altitudeMock);
+
+    t.true(model._minimumAltitude === 23000);
+    t.true(model._maximumAltitude === 23000);
 });
 
 ava('._calculateMaximumDelayFromSpawnRate() returns a number equal to 1hr in miliseconds / frequency', (t) => {